Move majority of context building to context builder
